Split notifications file into different files
